move terminal launcher to saperate file
