+++ core/main/src/main/java/com/rk/drawer/DrawerRegistries.kt
@@ -21,38 +21,3 @@ object AddProjectRegistry {
     val options = mutableStateListOf<AddProjectOption>()
 }
 
-object TerminalLauncher {
-    // A delegate function to launch the terminal activity dynamically
-    var handler: ((
-        activity: Activity,
-        sandbox: Boolean,
-        exe: String,
-        args: Array<String>,
-        id: String,
-        terminatePreviousSession: Boolean,
-        workingDir: String?,
-        env: Array<String>
-    ) -> Unit)? = null
-
-    fun launch(
-        activity: Activity,
-        sandbox: Boolean = true,
-        exe: String,
-        args: Array<String> = arrayOf(),
-        id: String,
-        terminatePreviousSession: Boolean = true,
-        workingDir: String? = null,
-        env: Array<String> = arrayOf()
-    ) {
-        handler?.invoke(activity, sandbox, exe, args, id, terminatePreviousSession, workingDir, env)
-    }
-}
-
-object SandboxedProcessRegistry {
-    // A delegate function to spawn a sandboxed PRoot process dynamically
-    var provider: (suspend (
-        command: List<String>,
-        workingDir: String?,
-        excludeMounts: List<String>
-    ) -> Process)? = null
-}
